backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.1
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2
